# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "WireGuard Mesh Manager API",
    "description": "Backend API for managing WireGuard VPN networks and devices",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/api-keys/api-keys/": {
      "get": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "List Api Keys",
        "description": "List API keys, optionally filtered by device ID.",
        "operationId": "list_api_keys_api_api_keys_api_keys__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/APIKeyResponse"
                  },
                  "title": "Response List Api Keys Api Api Keys Api Keys  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Create Api Key",
        "description": "Create a new API key for a device.",
        "operationId": "create_api_key_api_api_keys_api_keys__post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/api-keys/api-keys/device/{device_id}": {
      "get": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Get Api Keys By Device",
        "description": "Get all API keys for a specific device.",
        "operationId": "get_api_keys_by_device_api_api_keys_api_keys_device__device_id__get",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/APIKeyResponse"
                  },
                  "title": "Response Get Api Keys By Device Api Api Keys Api Keys Device  Device Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/api-keys/api-keys/{api_key_id}": {
      "get": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Get Api Key",
        "description": "Get a specific API key by ID.",
        "operationId": "get_api_key_api_api_keys_api_keys__api_key_id__get",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Api Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Update Api Key",
        "description": "Update an API key.",
        "operationId": "update_api_key_api_api_keys_api_keys__api_key_id__put",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Api Key Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/APIKeyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Delete Api Key",
        "description": "Delete an API key permanently.",
        "operationId": "delete_api_key_api_api_keys_api_keys__api_key_id__delete",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Api Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Delete Api Key Api Api Keys Api Keys  Api Key Id  Delete"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/api-keys/api-keys/{api_key_id}/rotate": {
      "post": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Rotate Api Key",
        "description": "Rotate an API key, creating a new key and invalidating the old one.",
        "operationId": "rotate_api_key_api_api_keys_api_keys__api_key_id__rotate_post",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Api Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyRotateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/api-keys/api-keys/device/{device_id}/regenerate": {
      "post": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Regenerate Device Api Key",
        "description": "Regenerate an API key for a device (rotate existing or create new).",
        "operationId": "regenerate_device_api_key_api_api_keys_api_keys_device__device_id__regenerate_post",
        "parameters": [
          {
            "name": "device_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Device Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/APIKeyCreateResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/api-keys/api-keys/{api_key_id}/revoke": {
      "post": {
        "tags": [
          "api-keys",
          "api-keys"
        ],
        "summary": "Revoke Api Key",
        "description": "Revoke (disable) an API key.",
        "operationId": "revoke_api_key_api_api_keys_api_keys__api_key_id__revoke_post",
        "parameters": [
          {
            "name": "api_key_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Api Key Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Revoke Api Key Api Api Keys Api Keys  Api Key Id  Revoke Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "APIKeyCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "allowed_ip_ranges": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Ip Ranges"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "device_id": {
            "type": "string",
            "title": "Device Id"
          }
        },
        "type": "object",
        "required": [
          "name",
          "device_id"
        ],
        "title": "APIKeyCreate",
        "description": "Schema for creating a new API key."
      },
      "APIKeyCreateResponse": {
        "properties": {
          "api_key": {
            "$ref": "#/components/schemas/APIKeyResponse"
          },
          "key_value": {
            "type": "string",
            "title": "Key Value"
          }
        },
        "type": "object",
        "required": [
          "api_key",
          "key_value"
        ],
        "title": "APIKeyCreateResponse",
        "description": "Schema for API key creation response."
      },
      "APIKeyResponse": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "allowed_ip_ranges": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Ip Ranges"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled",
            "default": true
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "device_id": {
            "type": "string",
            "title": "Device Id"
          },
          "network_id": {
            "type": "string",
            "title": "Network Id"
          },
          "last_used_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "name",
          "id",
          "device_id",
          "network_id",
          "created_at",
          "updated_at"
        ],
        "title": "APIKeyResponse",
        "description": "Schema for API key responses."
      },
      "APIKeyRotateResponse": {
        "properties": {
          "old_key": {
            "$ref": "#/components/schemas/APIKeyResponse"
          },
          "new_key": {
            "$ref": "#/components/schemas/APIKeyCreateResponse"
          }
        },
        "type": "object",
        "required": [
          "old_key",
          "new_key"
        ],
        "title": "APIKeyRotateResponse",
        "description": "Schema for API key rotation response."
      },
      "APIKeyUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "allowed_ip_ranges": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 1000,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Allowed Ip Ranges"
          },
          "enabled": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Enabled"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          }
        },
        "type": "object",
        "title": "APIKeyUpdate",
        "description": "Schema for updating an API key."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}